Change money vars to numeric

+++ queries/Proc.sql
@@ -45,7 +45,7 @@ CREATE OR REPLACE FUNCTION check_pledge_amt ()
 RETURNS TRIGGER AS $$ 
 
 DECLARE  
-	reward_min_amt INT;
+	reward_min_amt NUMERIC;
 
 BEGIN
 
@@ -186,8 +186,8 @@ RETURNS TRIGGER AS $$
 
 DECLARE  
 	deadline DATE;
-	pledged_amt INT;
-	funding_goal INT;
+	pledged_amt NUMERIC;
+	funding_goal NUMERIC;
 BEGIN
 
     SELECT Projects.deadline INTO deadline
@@ -424,7 +424,7 @@ DECLARE
     -- Projects where we have execeeded the goal at some point, grouped by id backing and amount
     FROM Projects P INNER JOIN Backs B ON P.id = B.id WHERE (SELECT SUM(amount) FROM Backs B WHERE B.id = P.id) >= P.goal AND P.created <= today GROUP BY P.id, B.backing, B.amount); 
     r RECORD;
-    funded INT := 0;
+    funded NUMERIC := 0;
     project_goal NUMERIC := 0;
     look_for_next BOOLEAN := FALSE;
     prev_proj_id INT := 0;
